refactor(dotcom): move ZeroLogBuffer to its own file
